requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
aiofiles>=24.1.0
PyMuPDF>=1.24.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
numpy>=1.26.0,<2.0.0
